server/python-agent/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2 